Move method to its own file so that it can be tested. Add a return.
