+++ nextFolderNameTask/index.ts
@@ -1,4 +1,5 @@
 import tl = require('azure-pipelines-task-lib/task');
+import { getDirItems } from './NextNameGenerator';
 
 async function run() {
     try {
@@ -12,39 +13,6 @@ async function run() {
     }
 }
 
-function getDirItems(folderPath :string, filterTerm: string)
-{
-    //requiring path and fs modules
-    const path = require('path');
-    const fs = require('fs');
 
-    //joining path of directory 
-    const directoryPath = path.join(folderPath);
-
-    //passing directoryPath and callback function
-    fs.readdir(directoryPath, function (err: string, files: string[]) {
-        //handling error
-        if (err) {
-            return console.log('Unable to scan directory: ' + err);
-        } 
-        //listing all files using forEach
-        files.forEach(function (file: string) {
-            // Do whatever you want to do with the file
-            console.log(file);
-        })
-
-        console.log("Filtering with term: " + filterTerm);
-        files.forEach(function (file: string) {
-            // Do whatever you want to do with the file
-            if (file.match('RC'))
-            {
-                var number = file.replace('RC', '');
-                var nInt = parseInt(number);
-                console.log(nInt);
-            }
-        }
-        );
-    });
-}
 
 run();
